Add MSK-ACCESS support for CVR pipeline
