+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/clinical/CVRClinicalDataWriter.java
@@ -36,13 +36,13 @@ import java.io.*;
 import java.util.*;
 import org.apache.commons.lang.StringUtils;
 import org.cbioportal.cmo.pipelines.cvr.CVRUtilities;
-import org.cbioportal.cmo.pipelines.cvr.model.CVRClinicalRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRClinicalRecord;
 import org.springframework.batch.item.*;
 import org.springframework.batch.item.file.*;
 import org.springframework.batch.item.file.transform.PassThroughLineAggregator;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.*;
-import org.cbioportal.cmo.pipelines.cvr.model.CompositeClinicalRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.composite.CompositeClinicalRecord;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
