+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/clinical/MskimpactAgeFieldSetMapper.java
@@ -34,7 +34,7 @@ package org.cbioportal.cmo.pipelines.cvr.clinical;
 
 import java.util.List;
 import org.apache.log4j.Logger;
-import org.cbioportal.cmo.pipelines.cvr.model.MskimpactAge;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.MskimpactAge;
 import org.springframework.batch.item.file.mapping.FieldSetMapper;
 import org.springframework.batch.item.file.transform.FieldSet;
 import org.springframework.validation.BindException;
