+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/fusion/CVRFusionFieldSetMapper.java
@@ -34,7 +34,7 @@ package org.cbioportal.cmo.pipelines.cvr.fusion;
 
 import java.util.List;
 import org.apache.log4j.Logger;
-import org.cbioportal.cmo.pipelines.cvr.model.CVRFusionRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRFusionRecord;
 import org.springframework.batch.item.file.mapping.FieldSetMapper;
 import org.springframework.batch.item.file.transform.FieldSet;
 import org.springframework.validation.BindException;
