+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/model/staging/CVRFusionRecord.java
@@ -30,12 +30,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package org.cbioportal.cmo.pipelines.cvr.model;
+package org.cbioportal.cmo.pipelines.cvr.model.staging;
 
 import com.mysql.jdbc.StringUtils;
 import java.util.ArrayList;
 import java.util.List;
 import joptsimple.internal.Strings;
+import org.cbioportal.cmo.pipelines.cvr.model.CVRSvVariant;
+import org.cbioportal.cmo.pipelines.cvr.model.GMLCnvIntragenicVariant;
 
 /**
  *
