+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/model/staging/CVRSvRecord.java
@@ -30,9 +30,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package org.cbioportal.cmo.pipelines.cvr.model;
+package org.cbioportal.cmo.pipelines.cvr.model.staging;
 
 import java.util.*;
+import org.cbioportal.cmo.pipelines.cvr.model.CVRSvVariant;
 import org.springframework.batch.core.configuration.annotation.StepScope;
 
 /**
