+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/model/staging/LinkedMskimpactCaseRecord.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.cbioportal.cmo.pipelines.cvr.model;
+package org.cbioportal.cmo.pipelines.cvr.model.staging;
 
 import java.util.*;
