+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/model/staging/MskimpactSeqDate.java
@@ -29,38 +29,66 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-
-package org.cbioportal.cmo.pipelines.cvr.model;
+package org.cbioportal.cmo.pipelines.cvr.model.staging;
 
 import java.util.*;
 
-public class CompositeSegRecord {
-    private String newSegRecord;
-    private String oldSegRecord;
+/**
+ *
+ * @author heinsz
+ */
+public class MskimpactSeqDate {
+    private String SAMPLE_ID;
+    private String PATIENT_ID;
+    private String SEQ_DATE;
 
-    public CompositeSegRecord() {
-        this.newSegRecord = "";
-        this.oldSegRecord = "";
+    /**
+     * @return the SAMPLE_ID
+     */
+    public String getSAMPLE_ID() {
+        return SAMPLE_ID != null ? SAMPLE_ID : "";
     }
 
-    public CompositeSegRecord(String newSegRecord, String oldSegRecord) {
-        this.newSegRecord = newSegRecord;
-        this.oldSegRecord = oldSegRecord;
+    /**
+     * @param SAMPLE_ID the SAMPLE_ID to set
+     */
+    public void setSAMPLE_ID(String SAMPLE_ID) {
+        this.SAMPLE_ID = SAMPLE_ID;
     }
 
-    public void setNewSegRecord(String newSegRecord) {
-        this.newSegRecord = newSegRecord;
+    /**
+     * @return the PATIENT_ID
+     */
+    public String getPATIENT_ID() {
+        return PATIENT_ID != null ? PATIENT_ID : "";
     }
 
-    public String getNewSegRecord() {
-        return this.newSegRecord != "" ? this.newSegRecord : null;
+    /**
+     * @param PATIENT_ID the PATIENT_ID to set
+     */
+    public void setPATIENT_ID(String PATIENT_ID) {
+        this.PATIENT_ID = PATIENT_ID;
     }
 
-    public void setOldSegRecord(String oldSegRecord) {
-        this.oldSegRecord = oldSegRecord;
+    /**
+     * @return the SEQ_DATE
+     */
+    public String getSEQ_DATE() {
+        return SEQ_DATE != null ? SEQ_DATE : "";
     }
 
-    public String getOldSegRecord() {
-        return this.oldSegRecord != "" ? this.oldSegRecord : null;
+    /**
+     * @param SEQ_DATE the SEQ_DATE to set
+     */
+    public void setSEQ_DATE(String SEQ_DATE) {
+        this.SEQ_DATE = SEQ_DATE;
+    }
+    
+    public static List<String> getFieldNames() {
+        List<String> fieldNames = new ArrayList<>();
+        fieldNames.add("SAMPLE_ID");
+        fieldNames.add("PATIENT_ID");
+        fieldNames.add("SEQ_DATE");
+        return fieldNames;
     }
 }
