+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/seg/CVRNewSegDataWriter.java
@@ -38,8 +38,8 @@ import java.io.Writer;
 import java.util.*;
 import org.apache.commons.lang.StringUtils;
 import org.cbioportal.cmo.pipelines.cvr.CVRUtilities;
-import org.cbioportal.cmo.pipelines.cvr.model.CVRSegRecord;
-import org.cbioportal.cmo.pipelines.cvr.model.CompositeSegRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRSegRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.composite.CompositeSegRecord;
 import org.springframework.batch.item.ExecutionContext;
 import org.springframework.batch.item.ItemStreamException;
 import org.springframework.batch.item.ItemStreamWriter;
