+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/seg/CVRSegDataProcessor.java
@@ -32,6 +32,8 @@
 
 package org.cbioportal.cmo.pipelines.cvr.seg;
 
+import org.cbioportal.cmo.pipelines.cvr.model.composite.CompositeSegRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRSegRecord;
 import java.util.*;
 import org.apache.commons.lang.StringUtils;
 import org.cbioportal.cmo.pipelines.cvr.CvrSampleListUtil;
