+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/seg/CVRSegDataWriter.java
@@ -32,6 +32,8 @@
 
 package org.cbioportal.cmo.pipelines.cvr.seg;
 
+import org.cbioportal.cmo.pipelines.cvr.model.composite.CompositeSegRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRSegRecord;
 import org.cbioportal.cmo.pipelines.cvr.CVRUtilities;
 import org.cbioportal.cmo.pipelines.cvr.model.*;
