+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/seg/CVRSegFieldSetMapper.java
@@ -39,7 +39,7 @@ package org.cbioportal.cmo.pipelines.cvr.seg;
 
 import java.util.List;
 import org.apache.log4j.Logger;
-import org.cbioportal.cmo.pipelines.cvr.model.CVRSegRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRSegRecord;
 import org.springframework.batch.item.file.mapping.FieldSetMapper;
 import org.springframework.batch.item.file.transform.FieldSet;
 import org.springframework.validation.BindException;
