+++ cvr/src/main/java/org/cbioportal/cmo/pipelines/cvr/sv/CVRSvDataWriter.java
@@ -38,8 +38,8 @@ import java.io.Writer;
 import java.util.*;
 import org.apache.commons.lang.StringUtils;
 import org.cbioportal.cmo.pipelines.cvr.CVRUtilities;
-import org.cbioportal.cmo.pipelines.cvr.model.CVRSvRecord;
-import org.cbioportal.cmo.pipelines.cvr.model.CompositeSvRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.staging.CVRSvRecord;
+import org.cbioportal.cmo.pipelines.cvr.model.composite.CompositeSvRecord;
 import org.springframework.batch.item.ExecutionContext;
 import org.springframework.batch.item.ItemStreamException;
 import org.springframework.batch.item.ItemStreamWriter;
